+++ history-parse/Processing/Page/Model/State/MatchModel.cs
@@ -1,7 +1,7 @@
-﻿using CSGOStats.Infrastructure.PageParse.Page.Parsing;
+﻿using CSGOStats.Infrastructure.PageParse.Extraction;
+using CSGOStats.Infrastructure.PageParse.Page.Parsing;
 using CSGOStats.Infrastructure.PageParse.Structure.Containers;
 using CSGOStats.Infrastructure.PageParse.Structure.Markers;
-using CSGOStats.Services.HistoryParse.Processing.Mappings.Attributes;
 
 namespace CSGOStats.Services.HistoryParse.Processing.Page.Model.State
 {
@@ -11,13 +11,13 @@ namespace CSGOStats.Services.HistoryParse.Processing.Page.Model.State
         [RequiredContainer("a"), AnchorLinkValue]
         public string Link { get; private set; }
 
-        [RequiredContainer("a/div/table/tr/td[@class = 'team-cell'][1]/div/img"), ImageTitleValue]
+        [RequiredContainer("a/div/table/tr/td[@class = 'team-cell'][1]/div/img"), ElementTitleValue]
         public string Team1 { get; private set; }
 
-        [RequiredContainer("a/div/table/tr/td[@class = 'team-cell'][2]/div/img"), ImageTitleValue]
+        [RequiredContainer("a/div/table/tr/td[@class = 'team-cell'][2]/div/img"), ElementTitleValue]
         public string Team2 { get; private set; }
 
-        [RequiredContainer("a/div/table/tr/td[@class = 'event']/img"), ImageTitleValue]
+        [RequiredContainer("a/div/table/tr/td[@class = 'event']/img"), ElementTitleValue]
         public string Event { get; private set; }
 
         [Collection, OptionalContainer("a/div/table/tr/td[@class = 'star-cell']/div[@class = 'map-and-stars']/div[@class = 'stars']/i")]